+++ tap_zendesk/__init__.py
@@ -1,18 +1,24 @@
 #!/usr/bin/env python3
 import json
 import sys
-import requests
-from requests.adapters import HTTPAdapter
+
 from zenpy import Zenpy
-from zenpy.lib.exception import ZenpyException
+import requests
+from requests import Session
+from requests.adapters import HTTPAdapter
+from requests.exceptions import Timeout, ChunkedEncodingError
+from urllib3.exceptions import ProtocolError
 import singer
-from singer import metadata
-from tap_zendesk.discover import discover_streams
-from tap_zendesk.sync import sync_stream
-from tap_zendesk.streams import STREAMS
+from singer import metadata, metrics as singer_metrics
+import backoff
 from tap_zendesk import metrics as zendesk_metrics
+from tap_zendesk.discover import discover_streams
+from tap_zendesk.streams import STREAMS
+from tap_zendesk.sync import sync_stream
 
 LOGGER = singer.get_logger()
+
+REQUEST_TIMEOUT = 300
 
 REQUIRED_CONFIG_KEYS = [
     "start_date",
@@ -30,14 +36,40 @@ API_TOKEN_CONFIG_KEYS = [
     "api_token",
 ]
 
-def do_discover(client):
+# patch Session.request to record HTTP request metrics
+request = Session.request
+
+
+@backoff.on_exception(backoff.expo,
+                      (ConnectionError, ConnectionResetError, Timeout, ChunkedEncodingError,
+                       ProtocolError),
+                      max_tries=5,
+                      factor=2)
+def request_metrics_patch(self, method, url, **kwargs):
+    with singer_metrics.http_request_timer(None):
+        response = request(self, method, url, **kwargs)
+        LOGGER.info("Request: %s, Response ETag: %s, Request Id: %s",
+                    url,
+                    response.headers.get('ETag', 'Not present'),
+                    response.headers.get('X-Request-Id', 'Not present'))
+        return response
+
+
+Session.request = request_metrics_patch
+
+
+# end patch
+
+def do_discover(client, config):
     LOGGER.info("Starting discover")
-    catalog = {"streams": discover_streams(client)}
+    catalog = {"streams": discover_streams(client, config)}
     json.dump(catalog, sys.stdout, indent=2)
     LOGGER.info("Finished discover")
 
+
 def stream_is_selected(mdata):
     return mdata.get((), {}).get('selected', False)
+
 
 def get_selected_streams(catalog):
     selected_stream_names = []
@@ -54,14 +86,17 @@ SUB_STREAMS = {
     'ticket_audits': ['ticket_audit_events'],
 }
 
+
 def get_sub_stream_names():
     sub_stream_names = []
     for parent_stream in SUB_STREAMS:
         sub_stream_names.extend(SUB_STREAMS[parent_stream])
     return sub_stream_names
 
+
 class DependencyException(Exception):
     pass
+
 
 def validate_dependencies(selected_stream_ids):
     errs = []
@@ -76,12 +111,14 @@ def validate_dependencies(selected_stream_ids):
     if errs:
         raise DependencyException(" ".join(errs))
 
+
 def populate_class_schemas(catalog, selected_stream_names):
     for stream in catalog.streams:
         if stream.tap_stream_id in selected_stream_names:
             STREAMS[stream.tap_stream_id].stream = stream
 
-def do_sync(client, config, catalog, state):
+
+def do_sync(client, catalog, state, config):
     selected_stream_names = get_selected_streams(catalog)
     validate_dependencies(selected_stream_names)
     populate_class_schemas(catalog, selected_stream_names)
@@ -106,7 +143,6 @@ def do_sync(client, config, catalog, state):
         # else:
         #     LOGGER.info("%s: Starting", stream_name)
 
-
         key_properties = metadata.get(mdata, (), 'table-key-properties')
         singer.write_schema(stream_name, stream.schema.to_dict(), key_properties)
 
@@ -118,7 +154,8 @@ def do_sync(client, config, catalog, state):
                 sub_stream = STREAMS[sub_stream_name].stream
                 sub_mdata = metadata.to_map(sub_stream.metadata)
                 sub_key_properties = metadata.get(sub_mdata, (), 'table-key-properties')
-                singer.write_schema(sub_stream.tap_stream_id, sub_stream.schema.to_dict(), sub_key_properties)
+                singer.write_schema(sub_stream.tap_stream_id, sub_stream.schema.to_dict(),
+                                    sub_key_properties)
 
         # parent stream will sync sub stream
         if stream_name in all_sub_stream_names:
@@ -126,22 +163,15 @@ def do_sync(client, config, catalog, state):
 
         LOGGER.info("%s: Starting sync", stream_name)
         instance = STREAMS[stream_name](client, config)
-        try:
-            counter_value = sync_stream(state, instance)
-            singer.write_state(state)
-            LOGGER.info("%s: Completed sync (%s rows)", stream_name, counter_value)
-            zendesk_metrics.log_aggregate_rates()
-        except Exception:
-            # Don't allow an error in one stream prevent other streams
-            # from running
-            sync_errors += 1
-            LOGGER.exception("%s: Sync failed", stream_name)
+        counter_value = sync_stream(state, config.get('start_date'), instance)
+        singer.write_state(state)
+        LOGGER.info("%s: Completed sync (%s rows)", stream_name, counter_value)
+        zendesk_metrics.log_aggregate_rates()
 
     singer.write_state(state)
     LOGGER.info("Finished sync")
     zendesk_metrics.log_aggregate_rates()
 
-    sys.exit(sync_errors)
 
 def oauth_auth(args):
     if not set(OAUTH_CONFIG_KEYS).issubset(args.config.keys()):
@@ -153,6 +183,7 @@ def oauth_auth(args):
         "subdomain": args.config['subdomain'],
         "oauth_token": args.config['access_token'],
     }
+
 
 def api_token_auth(args):
     if not set(API_TOKEN_CONFIG_KEYS).issubset(args.config.keys()):
@@ -166,6 +197,7 @@ def api_token_auth(args):
         "token": args.config['api_token']
     }
 
+
 def get_session(config):
     """ Add partner information to requests Session object if specified in the config. """
     if not all(k in config for k in ["marketplace_name",
@@ -177,36 +209,33 @@ def get_session(config):
     # https://github.com/facetoe/zenpy/blob/master/docs/zenpy.rst#usage
     session.mount("https://", HTTPAdapter(**Zenpy.http_adapter_kwargs()))
     session.headers["X-Zendesk-Marketplace-Name"] = config.get("marketplace_name", "")
-    session.headers["X-Zendesk-Marketplace-Organization-Id"] = str(config.get("marketplace_organization_id", ""))
+    session.headers["X-Zendesk-Marketplace-Organization-Id"] = str(
+        config.get("marketplace_organization_id", ""))
     session.headers["X-Zendesk-Marketplace-App-Id"] = str(config.get("marketplace_app_id", ""))
     return session
+
 
 @singer.utils.handle_top_exception(LOGGER)
 def main():
     parsed_args = singer.utils.parse_args(REQUIRED_CONFIG_KEYS)
 
+    # Set request timeout to config param `request_timeout` value.
+    config_request_timeout = parsed_args.config.get('request_timeout')
+    if config_request_timeout and float(config_request_timeout):
+        request_timeout = float(config_request_timeout)
+    else:
+        request_timeout = REQUEST_TIMEOUT  # If value is 0, "0", "" or not passed then it sets default to 300 seconds.
     # OAuth has precedence
     creds = oauth_auth(parsed_args) or api_token_auth(parsed_args)
     session = get_session(parsed_args.config)
-
-    # Pass some config options into the client
-    rate_limit_settings = {}
-    for key in ('proactive_ratelimit', 'proactive_ratelimit_request_interval', 'ratelimit_budget'):
-        if key in parsed_args.config:
-            rate_limit_settings[key] = parsed_args.config.pop(key)
-    LOGGER.info('zenpy rate limit settings = %r', rate_limit_settings)
-    client = Zenpy(session=session, **rate_limit_settings, **creds)
+    client = Zenpy(session=session, timeout=request_timeout, **creds)  # Pass request timeout
 
     if not client:
         LOGGER.error("""No suitable authentication keys provided.""")
 
     if parsed_args.discover:
-        do_discover(client)
+        # passing the config to check the authentication in the do_discover method
+        do_discover(client, parsed_args.config)
     elif parsed_args.catalog:
         state = parsed_args.state
-        
-        filtered_config = {
-            k: v for k, v in parsed_args.config.items()
-            if k not in OAUTH_CONFIG_KEYS + API_TOKEN_CONFIG_KEYS
-        }
-        do_sync(client, filtered_config, parsed_args.catalog, state)
+        do_sync(client, parsed_args.catalog, state, parsed_args.config)